hw-06-orm_jpa/src/main/java/ru/otus/homework/repositories/impl/CommentRepositoryJpa.java
package ru.otus.homework.repositories.impl;

import ru.otus.homework.entity.Comment;
import ru.otus.homework.repositories.CommentRepository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import java.util.List;
import java.util.Optional;

public class CommentRepositoryJpa implements CommentRepository {
    @PersistenceContext
    private final EntityManager entityManager;

    public CommentRepositoryJpa(EntityManager entityManager) {
        this.entityManager = entityManager;
    }

    @Override
    public Comment save(Comment comment) {
        if (comment.getId() == null) {
            entityManager.persist(comment);
            return comment;
        } else {
            return entityManager.merge(comment);
        }
    }

    @Override
    public Optional<Comment> findById(long id) {
        return Optional.ofNullable(entityManager.find(Comment.class, id));
    }

    @Override
    public List<Comment> findAll() {
        return entityManager.createQuery("select c from Comment c", Comment.class).getResultList();
    }

    @Override
    public void updateCommentById(long id, String commentText) {
        entityManager.createQuery("update Comment c set c.commentText = :commentText where c.id = :id")
                .setParameter("id", id)
                .setParameter("commentText", commentText)
                .executeUpdate();
    }

    @Override
    public void deleteById(long id) {
        entityManager.createQuery("delete from Comment c where c.id = :id")
                .setParameter("id", id)
                .executeUpdate();
    }
}
